requirements.txt
# Django & DRF
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# Celery & Redis
celery[redis]>=5.3.0
redis>=5.0.0

# Excel ingestion
pandas>=2.0.0
openpyxl>=3.1.0

# Production
gunicorn>=21.0.0
